Moved adapter utils classes to adapter package.
Issue: eclipse#426.

These classes mostly contain implementation details for adapter
implementations. With one exception
(SW360ComponentAdapterUtils.setComponentType()) they are only used
internally. The visibility of classes and methods was reduced as
possible.
