+++ modules/sw360/sw360-client/src/main/java/org/eclipse/sw360/antenna/sw360/client/adapter/SW360ProjectAdapterUtils.java
@@ -9,14 +9,14 @@
  *
  * SPDX-License-Identifier: EPL-2.0
  */
-package org.eclipse.sw360.antenna.sw360.utils;
+package org.eclipse.sw360.antenna.sw360.client.adapter;
 
 import org.apache.commons.lang3.StringUtils;
 import org.eclipse.sw360.antenna.sw360.rest.resource.SW360Visibility;
 import org.eclipse.sw360.antenna.sw360.rest.resource.projects.SW360Project;
 import org.eclipse.sw360.antenna.sw360.rest.resource.projects.SW360ProjectType;
 
-public class SW360ProjectAdapterUtils {
+class SW360ProjectAdapterUtils {
 
     private SW360ProjectAdapterUtils() {}
