+++ modules/sw360/sw360-client/src/main/java/org/eclipse/sw360/antenna/sw360/client/adapter/SW360ReleaseAdapterUtils.java
@@ -9,11 +9,11 @@
  *
  * SPDX-License-Identifier: EPL-2.0
  */
-package org.eclipse.sw360.antenna.sw360.utils;
+package org.eclipse.sw360.antenna.sw360.client.adapter;
 
 import org.eclipse.sw360.antenna.sw360.rest.resource.releases.SW360Release;
 
-public class SW360ReleaseAdapterUtils {
+class SW360ReleaseAdapterUtils {
 
     private SW360ReleaseAdapterUtils() {}
