+++ modules/sw360/sw360-client/src/test/java/org/eclipse/sw360/antenna/sw360/client/adapter/SW360ProjectClientAdapterAsyncImplTest.java
@@ -17,7 +17,6 @@ import org.eclipse.sw360.antenna.sw360.rest.resource.Self;
 import org.eclipse.sw360.antenna.sw360.rest.resource.projects.SW360Project;
 import org.eclipse.sw360.antenna.sw360.rest.resource.releases.SW360Release;
 import org.eclipse.sw360.antenna.sw360.rest.resource.releases.SW360ReleaseLinkObjects;
-import org.eclipse.sw360.antenna.sw360.utils.SW360ProjectAdapterUtils;
 import org.junit.Before;
 import org.junit.Test;
